resnet/requirements.txt
# Runtime dependencies for landmark-embedding
torch>=2.1
torchvision>=0.16
timm>=1.0.9
tqdm>=4.66
numpy>=1.24
Pillow>=10
scikit-learn>=1.3
kagglehub>=0.2
